remove redundant/incorrect code.
there is no need for an `initialized_` flag, because the controller manager
will call `starting` to indicate the first iteration of a control cycle.

The controller already does the correct thing in `starting`. If instead the
code in `starting` were removed, and the `initialized_` flag kept, then this
position controller would (incorrectly) jump to the previous set point if it is
stopped and started (for example, if you hit and release the run-stop), since
the `initialized_` flag was only being cleared in the constructor.