calibration: turn countdown_ into parameter
For smooth slow motions through the zero crossing it can
be helpful to increase the distance before passing through it from low to high.

As a longer default time changes the well-established startup initialization
to the worse (more collisions) it's better to turn it into a parameter without
changing the default.